extract inferExtraProperties
