+++ src/lang/exp/enrich.ts
@@ -1,6 +1,5 @@
 import * as Cores from "../core"
-import { evaluate } from "../core"
-import { Ctx, ctxToEnv } from "../ctx"
+import { Ctx } from "../ctx"
 import { ElaborationError } from "../errors"
 import * as Exps from "../exp"
 import { checkProperties, Exp, infer, Inferred } from "../exp"
@@ -41,35 +40,18 @@ function enrichWithoutInfer(ctx: Ctx, exp: Exp, type: Value): Inferred {
       assertClazzInCtx(ctx, type)
 
       const properties = checkProperties(ctx, exp.properties, type)
+      const names = Object.keys(properties)
 
       /**
          Extra properties are not checked,
          thus we require that they are infer-able.
       **/
 
-      const names = Object.keys(properties)
-      const extraInferred = Object.entries(exp.properties)
-        .filter(([name, exp]) => !names.includes(name))
-        .map(([name, exp]): [string, Exps.Inferred] => [name, infer(ctx, exp)])
-      const extraProperties = Object.fromEntries(
-        extraInferred.map(([name, inferred]) => [name, inferred.core]),
-      )
-
-      const extraTypedValues = Object.fromEntries(
-        extraInferred.map(([name, inferred]) => [
-          name,
-          {
-            type: inferred.type,
-            value: evaluate(ctxToEnv(ctx), inferred.core),
-          },
-        ]),
-      )
-
-      const extraClazz = Values.clazzFromTypedValues(extraTypedValues)
+      const extra = Exps.inferExtraProperties(ctx, exp.properties, names)
 
       return Inferred(
-        Values.prependFulfilledClazz(extraClazz, type),
-        Cores.Objekt({ ...properties, ...extraProperties }),
+        Values.prependFulfilledClazz(extra.clazz, type),
+        Cores.Objekt({ ...properties, ...extra.properties }),
       )
     }
