+++ src/lang/exp/index.ts
@@ -9,6 +9,7 @@ export * from "./enrich"
 export * from "./enrichOrCheck"
 export * from "./Exp"
 export * from "./infer"
+export * from "./inferExtraProperties"
 export * from "./inferProperties"
 export * from "./prepareProperties"
 export * from "./Span"
